feat(scan): 掃描頁改用長線(60D)鏡頭 — scoreLong 過濾/排序/顯示
掃描頁原本 universe 過濾、排序、卡片顯示分數一律用 scoreShort(5D)。但 score_validate
實證 edge 在 60D(高分→報酬 spread +6.3% 單調)、5D 接近雜訊(+0.8% 近乎平)；對長期
持有的使用者，5D 既無 edge 又不符持有週期。

改動：
- ScanFilterService.applySort/buildStockItems 加 horizon 參數(預設 short 向後相容)，
  long → scoreLong
- ScanNotifier 定死 _horizon = Horizon.long(universe/排序/顯示全用 60D)

背景：舊全域 horizon 開關已於 2026-06-19 被 3-tab Mode UI 取代、無 UI 可切
(selectedHorizonProvider 成孤兒永遠 short)。故掃描頁不再依賴該死 provider、直接用
有 edge 的 60D；同步移除對它的 read + listener。全 app 預設 60D(個股詳情/比較也卡 5D)
是另一個獨立決定(B2)，未動。

短線分支保留 + 測試覆蓋(service 兩分支 + 回歸保護)；87 個 scan 相關測試全綠。

+++ test/presentation/providers/scan_provider_test.dart
@@ -145,7 +145,7 @@ void main() {
     Map<String, List<DailyReasonEntry>>? reasons,
   }) {
     when(
-      () => mockAnalysisRepo.findLatestAnalyses(horizon: Horizon.short),
+      () => mockAnalysisRepo.findLatestAnalyses(horizon: Horizon.long),
     ).thenAnswer((_) async => (targetDate: testDate, analyses: analyses));
     when(
       () => mockDb.getReasonsBatch(any(), any()),
@@ -331,7 +331,7 @@ void main() {
 
     test('loadData handles error gracefully', () async {
       when(
-        () => mockAnalysisRepo.findLatestAnalyses(horizon: Horizon.short),
+        () => mockAnalysisRepo.findLatestAnalyses(horizon: Horizon.long),
       ).thenThrow(Exception('Database error'));
 
       final notifier = container.read(scanProvider.notifier);
